\name{print.dsa}
\alias{print.partDSA}
\alias{print.dsa}
\title{Print a partDSA Object}
\description{
These functions print \code{partDSA} and \code{dsa} objects.
}
\usage{
\method{print}{partDSA}(x, \dots)
\method{print}{dsa}(x, \dots)
}
\arguments{
\item{x}{
fitted model object of class \code{partDSA} or \code{dsa}.
These are the result of calling the \code{partDSA}
or \code{rss.dsa} functions.
}
\item{\dots}{unused.}
}
\keyword{utilities}
